Connected tiles and ship sets. Functional ship detection.
Tiles and ships above them are assigned same id numbers at ship spawning.
When a tile is hit, it changes color to gray if it is empty, red if there's a ship on it.  Tiles can only be clicked on once.
Once the whole ship is sunk, it reveals itself and the tiles around it (no-go zone) are marked as empty.
Tiles dispatch events to which ships subscribe.
The sunken ship dispatches an event to which its no-go zone subscribes.